# docs/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Risk Management System API",
        "version": "1.0.0",
        "description": "A comprehensive Risk Management Backend System for portfolio risk assessment, scoring, and analysis capabilities for individual investors. Designed for integration with MuleSoft.\n\n## Overview\nThis API provides:\n- **Risk Tolerance Assessment** - 15-question questionnaire processing and scoring\n- **Portfolio Risk Analysis** - VaR, Sharpe Ratio, Beta, Standard Deviation calculations\n- **Concentration Risk Detection** - Single position, sector, and top-5 holdings analysis\n- **Stress Testing** - Scenario analysis with historical and hypothetical scenarios\n- **Investment Suitability** - Portfolio vs risk profile alignment checks\n- **Recommendations** - Generated investment recommendations\n\n## Authentication\nThis is a demo API with **no authentication required**.\n\n## Base URLs\n- **Production:** https://risk-management-system-f04b786dc797.herokuapp.com/api/v1\n- **Local:** http://localhost:3002/api/v1\n\n## Data Note\nThis system uses an **in-memory datastore** that resets when the server restarts. It comes pre-loaded with 50 sample investors.",
        "contact": {
            "name": "API Support",
            "email": "support@riskmanagement.demo"
        },
        "license": {
            "name": "ISC",
            "url": "https://opensource.org/licenses/ISC"
        }
    },
    "servers": [
        {
...
        },
        {
...
    "paths": {
...
        "/scenarios": {
            "get": {
                "summary": "List all scenarios",
                "description": "Returns all stress test scenarios (historical, hypothetical, and regulatory)",
                "tags": [
                    "Scenarios"
                ],
                "responses": {
                    "200": {
                        "description": "List of all scenarios",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "scenarios": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object",
                                                        "properties": {
                                                            "scenarioId": {
                                                                "type": "string",
                                                                "example": "SCN-001"
                                                            },
                                                            "scenarioName": {
                                                                "type": "string",
                                                                "example": "2008 Financial Crisis"
                                                            },
                                                            "scenarioDescription": {
                                                                "type": "string"
                                                            },
                                                            "scenarioCategory": {
                                                                "type": "string",
                                                                "enum": [
                                                                    "HISTORICAL",
                                                                    "HYPOTHETICAL",
                                                                    "REGULATORY"
                                                                ]
                                                            },
                                                            "isActive": {
                                                                "type": "boolean"
                                                            },
                                                            "shockParameters": {
                                                                "type": "object",
                                                                "properties": {
                                                                    "equityShock": {
                                                                        "type": "number"
                                                                    },
                                                                    "bondShock": {
                                                                        "type": "number"
                                                                    },
                                                                    "creditSpreadChange": {
                                                                        "type": "number"
                                                                    },
                                                                    "volatilitySpike": {
                                                                        "type": "number"
                                                                    }
                                                                }
                                                            },
                                                            "sectorShocks": {
                                                                "type": "object"
                                                            }
                                                        }
                                                    }
                                                },
                                                "count": {
                                                    "type": "integer",
                                                    "example": 10
                                                }
                                            }
                                        },
                                        "timestamp": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Create custom scenario",
                "description": "Creates a new stress test scenario",
                "tags": [
                    "Scenarios"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "scenarioName",
                                    "scenarioCategory",
                                    "shockParameters"
                                ],
                                "properties": {
                                    "scenarioName": {
                                        "type": "string",
                                        "example": "Custom Recession Scenario"
                                    },
                                    "scenarioDescription": {
                                        "type": "string",
                                        "example": "A custom stress scenario for testing"
                                    },
                                    "scenarioCategory": {
                                        "type": "string",
                                        "enum": [
                                            "HISTORICAL",
                                            "HYPOTHETICAL",
                                            "REGULATORY"
                                        ],
                                        "example": "HYPOTHETICAL"
                                    },
                                    "isActive": {
                                        "type": "boolean",
                                        "default": true
                                    },
                                    "shockParameters": {
                                        "type": "object",
                                        "properties": {
                                            "equityShock": {
                                                "type": "number",
                                                "example": -30
                                            },
                                            "bondShock": {
                                                "type": "number",
                                                "example": 5
                                            },
                                            "creditSpreadChange": {
                                                "type": "number",
                                                "example": 200
                                            },
                                            "volatilitySpike": {
                                                "type": "number",
                                                "example": 50
                                            }
                                        }
                                    },
                                    "sectorShocks": {
                                        "type": "object",
                                        "example": {
                                            "TECHNOLOGY": -35,
                                            "FINANCIAL": -40,
                                            "HEALTHCARE": -20
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Scenario created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "scenarioId": {
                                                    "type": "string"
                                                },
                                                "scenarioName": {
                                                    "type": "string"
                                                },
                                                "scenarioCategory": {
                                                    "type": "string"
                                                }
                                            }
                                        },
                                        "timestamp": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/scenarios/category/{category}": {
            "get": {
                "summary": "Get scenarios by category",
                "description": "Returns all scenarios of a specific category",
                "tags": [
                    "Scenarios"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "category",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "HISTORICAL",
                                "HYPOTHETICAL",
                                "REGULATORY"
                            ]
                        },
                        "description": "The scenario category",
                        "example": "HISTORICAL"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Scenarios retrieved",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "scenarios": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                },
                                                "count": {
                                                    "type": "integer"
                                                }
                                            }
                                        },
                                        "timestamp": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}